+++ packages/terra-table/src/subcomponents/ColumnHeaderCell.jsx
@@ -338,7 +338,7 @@ const ColumnHeaderCell = (props) => {
   // Add column highlight indicator based on color
   let columnHighlightIcon;
   // Column highlighting is supported for single column spans only
-  if (columnSpan == 1){
+  if (columnSpan === 1) {
     if (columnHighlightColor === ColumnHighlightColor.GREEN) {
       columnHighlightIcon = <svg className={cx('highlight-icon-svg')} xmlns="http://www.w3.org/2000/svg"><circle className={cx('highlight-icon-circle')} r="3" cx="110%" cy="11" transform="translate(-5)" /></svg>;
     } else if (columnHighlightColor === ColumnHighlightColor.ORANGE) {