+++ packages/terra-table/src/subcomponents/Row.jsx
@@ -139,7 +139,7 @@ function Row(props) {
   const heightProperties = (height) ? {
     height,
   } : { minHeight: rowMinimumHeight };
-  
+
   return (
     <tr
       aria-rowindex={rowIndex}
@@ -174,8 +174,8 @@ function Row(props) {
       {cells.map((cellData, index) => {
         const cellColumnIndex = index + columnIndexOffSet;
         const columnId = displayedColumns[cellColumnIndex].id;
-        const columnSpan = displayedColumns[cellColumnIndex].columnSpan;
-        
+        const { columnSpan } = displayedColumns[cellColumnIndex];
+
         return (
           <Cell
             rowId={id}